Simplification of the pairing
Realized a shortcut could be made. Instead of always putting people in the queue,
only do this when the queue is empty. This is avery naïve way of doing it, but
should not pose a problem with only a few users. A whole lot easier to demo which
is the inteded purpose of the app.